requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
ollama==0.1.6
pypdf==4.0.1
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2
send2trash==1.8.3
